+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/event_subscribers.xml
@@ -35,7 +35,7 @@
         </service>
 
         <service id="Sylius\Bundle\ApiBundle\EventSubscriber\PaymentRequestEventSubscriber">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncerInterface" />
+            <argument type="service" id="sylius.announcer.payment_request" />
             <tag name="kernel.event_subscriber" />
         </service>
