+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/order_pay/providers.xml
@@ -31,7 +31,7 @@
         <service abstract="true" id="Sylius\Bundle\CoreBundle\OrderPay\Provider\PaymentRequestPayResponseProvider" class="Sylius\Bundle\CoreBundle\OrderPay\Provider\PaymentRequestPayResponseProvider">
             <argument type="service" id="sylius.factory.payment_request" />
             <argument type="service" id="sylius.repository.payment_request" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncerInterface" />
+            <argument type="service" id="sylius.announcer.payment_request" />
             <argument type="service" id="sylius.payment_request.provider.http_response.default" />
             <argument type="service" id="sylius.payment_request.provider.default_action" />
             <argument type="service" id="sylius.payment_request.provider.default_payload" />
@@ -50,7 +50,7 @@
 
         <service abstract="true" id="Sylius\Bundle\CoreBundle\OrderPay\Provider\PaymentRequestAfterPayResponseProvider" class="Sylius\Bundle\CoreBundle\OrderPay\Provider\PaymentRequestAfterPayResponseProvider">
             <argument type="service" id="sylius.factory.payment_request" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncerInterface" />
+            <argument type="service" id="sylius.announcer.payment_request" />
             <argument type="service" id="sylius.payment_request.provider.http_response.default" />
             <argument type="service" id="sylius.repository.payment_request" />
             <argument type="service" id="sylius.payment_request.command_provider.default" />
