+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/payum/processors.xml
@@ -21,7 +21,7 @@
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessorInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\AfterTokenRequestProcessor">
             <argument type="service" id="sylius.factory.payment_request" />
             <argument type="service" id="sylius.repository.payment_request" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncerInterface" />
+            <argument type="service" id="sylius.announcer.payment_request" />
         </service>
 
         <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessorInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Payum\RequestProcessor">
