+++ src/Sylius/Bundle/PaymentBundle/Resources/config/services/announcer.xml
@@ -11,17 +11,18 @@
 
 -->
 
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
     <services>
         <defaults public="true" />
 
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncerInterface" class="Sylius\Bundle\CoreBundle\PaymentRequest\Announcer\PaymentRequestAnnouncer">
+        <service id="sylius.announcer.payment_request" class="Sylius\Bundle\PaymentBundle\Announcer\PaymentRequestAnnouncer">
             <argument type="service" id="sylius.payment_request.command_provider.default" />
             <argument type="service" id="sylius.payment_request.command_bus" />
         </service>
-
+        <service id="Sylius\Bundle\PaymentBundle\Announcer\PaymentRequestAnnouncerInterface" alias="sylius.announcer.payment_request" />
     </services>
 </container>
